Consolidate client types into one file
