--- src/client/config.ts
@@ -1,1 +1,0 @@
-export const CHANNEL_NAME = 'of-notifications-service-v1';
