+++ src/provider/model/Entity.ts
