+++ src/provider/persistence/dataLayer/repositories/RepositoryFactory.ts
@@ -1,4 +1,4 @@
-import {Entity} from '../../../../client/Entity';
+import {Entity} from '../../../model/Entity';
 import {Settings} from '../../../model/Settings';
 import {IDatastore} from '../IDatastore';
 import {IndexedDb} from '../IndexedDb';
